Add routes to landing page and map page
