+++ src/Pages/LandingPage/LandingPage.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class LandingPage extends Component {
+  state = {}
+  render() {
+    return (
+      <div>
+        Placeholder for LandingPage
+      </div>
+    );
+  }
+}
+
+export default LandingPage;
